requirements.txt
absl-py==2.3.1
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
arm_pytorch_utilities==0.4.3
asttokens==3.0.1
attrs==25.4.0
av==15.1.0
bddl==1.0.1
beaker-gantry==3.2.0
beaker-py==2.5.1
cachetools==6.2.4
-e git+ssh://git@github.com/Treeeplanter/calvin.git@de70e95649c6eb8ae32bc99f78f55f1ed69e234e#egg=Calvin&subdirectory=../../../third_party/calvin/calvin_models
-e git+ssh://git@github.com/Treeeplanter/calvin_env.git@bdb0e94b8710dd5b4ec7396ee2ce616c444a8d1b#egg=calvin_env&subdirectory=../../../../../third_party/calvin/calvin_env
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-help-colors==0.9.4
click-option-group==0.5.9
cloudpickle==3.1.2
cmake==4.1.3
colorlog==6.10.1
contourpy==1.3.3
cycler==0.12.1
dacite==1.9.2
dataclass-extensions==0.2.3
datasets==4.1.1
decorator==5.2.1
deepdiff==8.6.1
diffusers==0.35.2
dill==0.3.8
distro==1.9.0
docstring_parser==0.17.0
draccus==0.10.0
easydict==1.9
einops==0.8.1
etils==1.13.0
evdev==1.9.2
executing==2.2.1
Farama-Notifications==0.0.4
fast_kinematics==0.2.2
fastjsonschema==2.21.2
filelock==3.20.0
fonttools==4.60.1
freetype-py==2.5.1
frozenlist==1.8.0
fsspec==2025.3.0
future==0.18.2
gitdb==4.0.12
GitPython==3.1.45
glfw==2.10.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-crc32c==1.7.1
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.75.1
grpcio-status==1.71.2
gym==0.26.2
gym-notices==0.1.0
gymnasium==1.2.2
h11==0.16.0
h5py==3.15.1
hf-xet==1.1.10
hf_transfer==0.1.9
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.35.3
hydra-colorlog==1.2.0
hydra-core==1.3.2
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
inquirerpy==0.3.4
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonlines==4.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
jupytext==1.18.1
kiwisolver==1.4.9
kmeans-pytorch==0.3
-e git+https://github.com/Treeeplanter/lerobot.git@fbf0fb1e6a39a9223b61fcd09aead75c95363194#egg=lerobot&subdirectory=../../../third_party/lerobot
-e git+ssh://git@github.com/Treeeplanter/LIBERO-PRO.git@6adba85c03663a7dac0df22ae54dbcf069f55623#egg=libero&subdirectory=../../../third_party/libero_pro
llvmlite==0.45.1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdit-py-plugins==0.5.0
mdurl==0.1.2
mergedeep==1.3.4
mplib==0.1.1
mpmath==1.3.0
mujoco==3.4.0
multidict==6.7.0
multiprocess==0.70.16
mypy_extensions==1.1.0
nbformat==5.10.4
networkx==3.6
numba==0.62.1
numpy==1.26.4
numpy-quaternion==2024.0.13
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-ml-py==13.580.82
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
openai==2.8.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
orderly-set==5.5.0
packaging==25.0
pandas==2.3.3
parse==1.20.2
parso==0.8.5
petname==2.6
pexpect==4.9.0
pfzy==0.3.4
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybullet==3.2.7
pycollada==0.6
pydantic==2.12.2
pydantic_core==2.41.4
pyglet==2.1.11
Pygments==2.19.2
pynput==1.8.1
PyOpenGL==3.1.10
pyparsing==3.2.5
pyperclip==1.11.0
pyrender==0.1.45
pyserial==3.5
pytest==9.0.2
python-dateutil==2.9.0.post0
python-xlib==0.33
pytorch-kinematics==0.7.6
pytorch-seed==0.2.0
pytz==2025.2
PyYAML==6.0.3
pyyaml-include==1.4.1
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
rerun-sdk==0.26.2
rich==14.2.0
robosuite==1.4.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.6.2
sapien==3.0.1
scikit-learn==1.7.2
scipy==1.16.3
sentry-sdk==2.42.0
setuptools==80.9.0
shtab==1.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
-e git+ssh://git@github.com/Vision-Language-Steering/code.git@41d250c730e34b2f78f4621be7342e9fcfd8ae20#egg=VLS
sympy==1.14.0
tabulate==0.9.0
-e git+https://github.com/lukashermann/tacto.git@dd53360d9a8c186f0d6439372ec0be0fa5e21731#egg=tacto&subdirectory=../../../../../../../third_party/calvin/calvin_env/tacto
termcolor==3.1.0
threadpoolctl==3.6.0
timm==1.0.20
tokenizers==0.21.4
toml==0.10.2
tomli==2.3.0
toppra==0.6.3
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchcodec==0.5
torchvision==0.22.1+cu118
tqdm==4.67.1
traitlets==5.14.3
transformers @ git+https://github.com/huggingface/transformers.git@dcddb970176382c0fcf4521b0c0e6fc15894dfe0
transforms3d==0.4.2
trimesh==4.10.0
triton==3.3.1
typeguard==4.4.4
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tyro==0.9.35
tzdata==2025.2
urdfpy==0.0.22
uritemplate==4.2.0
urllib3==2.5.0
wandb==0.21.4
wcwidth==0.2.14
wheel==0.45.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
